UPDATE: 	1. make new header files for arranging functions and minishell 	structure. 	2. take apart some functions from integrated files
