+++ includes/mini_external.h
@@ -1,0 +1,35 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   mini_external.h                                    :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: haryu <marvin@42.fr>                       +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/06/08 14:34:28 by haryu             #+#    #+#             */
+/*   Updated: 2022/06/08 14:36:39 by haryu            ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#ifndef MINI_EXTERNAL_H
+# define MINI_EXTERNAL_H
+
+# include <stdio.h>
+# include <string.h>
+# include <stdlib.h>
+# include <unistd.h>
+# include <dirent.h>
+# include <fcntl.h>
+# include <errno.h>
+# include <signal.h>
+# include <sys/stat.h>
+# include <sys/wait.h>
+# include <sys/resource.h>
+# include <sys/types.h>
+# include <sys/ioctl.h>
+# include "./readline/readline.h"
+# include "./readline/history.h"
+# include <termios.h>
+# include <termcap.h>
+# include <term.h>
+
+#endif
