+++ includes/parse.h
@@ -1,24 +1,28 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   heredoc.c                                          :+:      :+:    :+:   */
+/*   parse.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: haryu <haryu@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/06/04 22:25:15 by haryu             #+#    #+#             */
-/*   Updated: 2022/06/08 16:52:13 by haryu            ###   ########.fr       */
+/*   Created: 2022/06/08 14:58:19 by haryu             #+#    #+#             */
+/*   Updated: 2022/06/08 16:49:18 by haryu            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../includes/minishell.h"
+#ifndef PARSE_H
+# define PARSE_H
 
-int heredoc(char **chunks, int height)
-{
-	int i;
+/* --------------------------------------------------------------------------*/
+/**
+ * @Synopsis		: split command line for making chunks.
+ *
+ * @Param line		: whole command line for making chunks.
+ *
+ * @Returns   		: chunks or NULL
+ */
+/* ----------------------------------------------------------------------------*/
+char	**vertical_split(char *line);
 
-	i = 0;
-	(void)chunks;
-	(void)height;
+#endif
 
-	return (FALSE);
-}
