+++ src/check_height.c
@@ -1,24 +1,40 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   heredoc.c                                          :+:      :+:    :+:   */
+/*   check_height.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: haryu <haryu@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/06/04 22:25:15 by haryu             #+#    #+#             */
-/*   Updated: 2022/06/08 16:52:13 by haryu            ###   ########.fr       */
+/*   Created: 2022/06/08 15:41:40 by haryu             #+#    #+#             */
+/*   Updated: 2022/06/08 15:52:52 by haryu            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/minishell.h"
 
-int heredoc(char **chunks, int height)
+size_t check_height(char *line)
 {
-	int i;
+	size_t ret;
+	int	sq;
+	int	dq;
+	int	i;
 
-	i = 0;
-	(void)chunks;
-	(void)height;
-
-	return (FALSE);
+	i = -1;
+	ret = 1;
+	sq = 0;
+	dq = 0;
+	while(line[++i])
+	{
+		if (line[i] == '\'' && sq == 0)
+			sq++;
+		else if (line[i] == '\'' && sq == 1)
+			sq--;
+		if (line[i] == '"' && dq == 0)
+			dq++;
+		else if (line[i] == '"' && dq == 1)
+			dq--;
+		if (line[i] == '|' && dq == 0 && sq == 0)
+			ret++;
+	}
+	return (ret);
 }
